parsing/requirements.txt
certifi==2017.4.17
chardet==3.0.4
click==6.7
Django==1.11.2
Flask==0.12.2
idna==2.5
image==1.5.5
imgurpython==1.1.7
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
numpy==1.13.0+mkl
olefile==0.44
opencv-python==3.2.0.7
Pillow==4.1.1
pytz==2017.2
requests==2.18.1
urllib3==1.21.1
Werkzeug==0.12.2
